refactor(utils): finish CheckReservedTags extraction
The move left utils/tags.go without its imports, both callers without
theirs, and scaleway calling the old unexported name, so nothing in the
tree compiled. The two provider test files still exercised the removed
package-local helper.

Test the helper once where it now lives, against its own prefix rather
than engine.LabelPrefix, since the point of the move was to make it
independent of the autoscaler's namespace.
